requirements.txt
tkinter
pyzkaccess
ctypes
pillow 